+++ exonum/src/events/codec.rs
@@ -111,7 +111,6 @@ mod test {
     use crypto::{gen_keypair_from_seed, Seed, SEED_LENGTH};
     use events::noise::{HandshakeParams, NoiseWrapper};
     use messages::{MessageBuffer, RawMessage};
-    use node::state::SharedConnectList;
 
     #[test]
     fn decode_message_valid_header_size() {
@@ -152,13 +151,7 @@ mod test {
     fn create_encrypted_codecs() -> (MessagesCodec, MessagesCodec) {
         let (public_key, secret_key) = gen_keypair_from_seed(&Seed::new([1; SEED_LENGTH]));
         let address: SocketAddr = "127.0.0.1:8000".parse().unwrap();
-        let mut params = HandshakeParams::new(
-            public_key,
-            secret_key,
-            SharedConnectList::default(),
-            1024,
-            address,
-        );
+        let mut params = HandshakeParams::with_default_params(public_key, secret_key, address);
         params.set_remote_key(public_key);
 
         let mut initiator = NoiseWrapper::initiator(&params).session;