+++ exonum/src/sandbox/sandbox.rs
@@ -38,7 +38,7 @@ use events::{
     NetworkEvent, NetworkRequest, TimeoutRequest,
 };
 use helpers::{Height, Milliseconds, Round, ValidatorId};
-use messages::{Any, Message, PeersExchange, PeersRequest, RawMessage, RawTransaction, Status};
+use messages::{Any, Message, PeerList, PeersRequest, RawMessage, RawTransaction, Status};
 use node::ConnectInfo;
 use node::{
     ApiSender, Configuration, ConnectList, ConnectListConfig, ExternalMessage, ListenerConfig,
@@ -1011,7 +1011,7 @@ mod tests {
         let (receiver_pk, receiver_sk) = (&s.p(ValidatorId(0)), &s.s(ValidatorId(0)));
 
         let peers_request = PeersRequest::new(sender_pk, receiver_pk, sender_sk);
-        let peers_response = PeersExchange::new(receiver_pk, &sender_pk, s.peers(), receiver_sk);
+        let peers_response = PeerList::new(receiver_pk, &sender_pk, s.peers(), receiver_sk);
 
         s.recv(&peers_request);
         s.send(s.a(ValidatorId(1)), &peers_response);
@@ -1052,7 +1052,7 @@ mod tests {
         let (receiver_pk, receiver_sk) = (&s.p(ValidatorId(0)), &s.s(ValidatorId(0)));
 
         let peers_request = PeersRequest::new(sender_pk, receiver_pk, sender_sk);
-        let peers_response = PeersExchange::new(receiver_pk, &sender_pk, vec![], receiver_sk);
+        let peers_response = PeerList::new(receiver_pk, &sender_pk, vec![], receiver_sk);
 
         s.recv(&peers_request);
         s.send(s.a(ValidatorId(1)), &peers_response);